+++ src/fitter/__init__.py
@@ -1,12 +1,13 @@
 from .FittersChain import FittersChain
 from .ModelFitter import ModelFitter
-from .NelderMeadFitter import NelderMeadFitter
-from .GibbsSamplerFitter import GibbsSamplerFitter
 from .BruteForceFitter import BruteForceFitter
 from .BGDFitter import BGDFitter
-from .MonteCarloFitter import MonteCarloFitter
+from .CoordinateDescentFitter import CoordinateDescentFitter
 
-__all__ = ['FittersChain', 'ModelFitter',
-           'NelderMeadFitter', 'GibbsSamplerFitter',
-           'BruteForceFitter', 'BGDFitter',
-           'MonteCarloFitter']
+__all__ = [
+    'FittersChain',
+    'ModelFitter',
+    'BruteForceFitter',
+    'BGDFitter',
+    'CoordinateDescentFitter',
+]